add auth to casefile and consultation routes

+++ api/src/routers/consultationRouter.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const router = new express.Router();
 const Consultation = require('../models/consultation.model');
+const auth = require('../middleware/auth');
 
-router.post('/consultations', async (req, res) => {
+router.post('/consultations', auth, async (req, res) => {
   const consultation = new Consultation(req.body);
 
   try {
@@ -18,7 +19,7 @@ router.post('/consultations', async (req, res) => {
 // GET /consultations?practitionerId=125689
 // GET /consultations?limit=10&skip=10
 // GET /consultations?sortBy=createdAt:desc
-router.get('/consultations', async (req, res) => {
+router.get('/consultations', auth, async (req, res) => {
   const match = {};
   const sort = {};
 
@@ -45,7 +46,7 @@ router.get('/consultations', async (req, res) => {
   }
 });
 
-router.get('/consultations/:id', async (req, res) => {
+router.get('/consultations/:id', auth, async (req, res) => {
   try {
     const consultation = await Consultation.findOne({ _id: req.params.id }).populate(
       'practitioner',
@@ -63,7 +64,7 @@ router.get('/consultations/:id', async (req, res) => {
   }
 });
 
-router.patch('/consultations/:id', async (req, res) => {
+router.patch('/consultations/:id', auth, async (req, res) => {
   const updates = Object.keys(req.body);
   const allowedUpdates = [
     'patientId',
@@ -96,7 +97,11 @@ router.patch('/consultations/:id', async (req, res) => {
   }
 });
 
-router.delete('/consultations/:id', async (req, res) => {
+router.delete('/consultations/:id', auth, async (req, res) => {
+  if (req.practitioner.accessLevel < 2) {
+    return res.status(403).send({ error: 'Forbidden to delete consultation' });
+  }
+
   try {
     const consultation = await Consultation.findOneAndDelete({ _id: req.params.id });
     if (!consultation) {
